用户修改接口开发

+++ shiguangji-user/shiguangji-user-biz/src/main/java/com/xiaowei/shiguangji/user/biz/domain/dataobject/UserDO.java
@@ -30,9 +30,9 @@ public class UserDO {
 
     private String phone;
 
-    private Byte sex;
+    private Integer sex;
 
-    private Byte status;
+    private Integer status;
 
     private String introduction;
